// src/nebular-react-styles/src/theme/types/nebular-theme/NebularThemeStepper.ts
export interface NebularThemeStepper {
  stepper_step_text_color: any;
  stepper_step_text_font_family: any;
  stepper_step_text_font_size: any;
  stepper_step_text_font_weight: any;
  stepper_step_text_line_height: any;
  stepper_step_active_text_color: any;
  stepper_step_completed_text_color: any;

  stepper_step_index_border_color: any;
  stepper_step_index_border_style: any;
  stepper_step_index_border_width: any;
  stepper_step_index_border_radius: any;
  stepper_step_index_width: any;
  stepper_step_index_active_border_color: any;
  stepper_step_index_completed_background_color: any;
  stepper_step_index_completed_border_color: any;
  stepper_step_index_completed_text_color: any;

  stepper_connector_background_color: any;
  stepper_connector_completed_background_color: any;
  stepper_horizontal_connector_margin: any;
  stepper_vertical_connector_margin: any;

  stepper_step_content_padding: any;
}